+++ e2e-tests/test_sora_mode_simulcast.py
@@ -11,10 +11,6 @@ pytestmark = [
         not os.environ.get("TEST_SORA_MODE_SIGNALING_URLS"),
         reason="TEST_SORA_MODE_SIGNALING_URLS not set in environment",
     ),
-    # pytest.mark.skipif(
-    #     platform.system() == "Darwin",
-    #     reason="Skipping simulcast test on macOS due to performance limitations",
-    # ),
 ]