+++ src/ayame/webrtc.rs
@@ -106,7 +106,7 @@ impl AyameEngine {
                     config.video_width,
                     config.video_height,
                     config.framerate,
-                    #[cfg(feature = "preview")]
+                    #[cfg(feature = "player")]
                     None,
                 );
                 info!(target: "ayame", width = config.video_width, height = config.video_height, fps = config.framerate, "fake video started");